Using convolutional LSTM instead of plain convolution
Results are roughly the same; maybe a bit blurrier. Hypothesis is
that this will capture temporal information better, in that state
will be retained further back in time than the fixed window that
was being used with the convlution.

Currently slower than the plain convolutional one, but may be
faster overall if we switch to stateful mode.